third_party/WebKit/LayoutTests/editing/deleting/merge-paragraph-into-blockquote.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        [
            '<div contenteditable>',
                '<blockquote>hello</blockquote> |world',
            '</div>',
        ].join(''),
        'delete',
        [
            '<div contenteditable>',
                '<blockquote>hello|world</blockquote>',
            '</div>',
        ].join('')),
        'Delete to merge paragraph into BLOCKQUOTE');
</script>
